+++ MultiFunPlayer/MediaSource/ViewModels/WhirligigMediaSourceViewModel.cs
@@ -77,7 +77,7 @@ public class WhirligigMediaSourceViewModel : AbstractMediaSource
                 else if (message.Length >= 8 && message[..8] == "duration")
                 {
                     var parts = message.Split('=', 2, StringSplitOptions.TrimEntries);
-                    if (parts.Length == 2 && float.TryParse(parts[1].Replace(',', '.'), NumberStyles.Any, NumberFormatInfo.InvariantInfo, out var duration) && duration >= 0)
+                    if (parts.Length == 2 && double.TryParse(parts[1].Replace(',', '.'), NumberStyles.Any, NumberFormatInfo.InvariantInfo, out var duration) && duration >= 0)
                         _eventAggregator.Publish(new MediaDurationChangedMessage(TimeSpan.FromSeconds(duration)));
                 }
                 else if (message.Length >= 1 && message[0] == 'P')
