+++ MultiFunPlayer/MotionProvider/AbstractMotionProvider.cs
@@ -13,11 +13,11 @@ public abstract class AbstractMotionProvider : Screen, IMotionProvider
     private readonly IEventAggregator _eventAggregator;
 
     public string Name => GetType().GetCustomAttribute<DisplayNameAttribute>(inherit: false).DisplayName;
-    public float Value { get; protected set; }
+    public double Value { get; protected set; }
 
-    [JsonProperty] public float Speed { get; set; } = 1;
-    [JsonProperty] public float Minimum { get; set; } = 0;
-    [JsonProperty] public float Maximum { get; set; } = 100;
+    [JsonProperty] public double Speed { get; set; } = 1;
+    [JsonProperty] public double Minimum { get; set; } = 0;
+    [JsonProperty] public double Maximum { get; set; } = 100;
 
     protected AbstractMotionProvider(DeviceAxis target, IEventAggregator eventAggregator)
     {
@@ -33,5 +33,5 @@ public abstract class AbstractMotionProvider : Screen, IMotionProvider
 
     protected virtual bool ShouldRequestSyncOnPropertyChange(string propertyName) => propertyName != nameof(Value);
 
-    public abstract void Update(float deltaTime);
+    public abstract void Update(double deltaTime);
 }
