+++ MultiFunPlayer/UI/Controls/DraggablePointCanvas.xaml.cs
@@ -137,8 +137,8 @@ public partial class DraggablePointCanvas : Canvas, INotifyPropertyChanged
             return;
 
         var position = e.GetPosition(this) - _captureOffset;
-        position.X = MathUtils.Clamp((float)position.X, 0, (float)ActualWidth);
-        position.Y = MathUtils.Clamp((float)position.Y, 0, (float)ActualHeight);
+        position.X = MathUtils.Clamp(position.X, 0, ActualWidth);
+        position.Y = MathUtils.Clamp(position.Y, 0, ActualHeight);
 
         point.Position = position;
         SynchronizePopup(position);
